Moved camera permsision to chat page
