+++ .gitignore
@@ -34,3 +34,6 @@ yarn-error.*
 
 # typescript
 *.tsbuildinfo
+
+/frontend/ios
+/frontend/android
